refactor(header): update navigation links to use "#" as href
The href values for "Blog" and "About" navigation items were updated to "#" to temporarily disable their functionality. This change is made to prevent broken links while the corresponding pages are under development or maintenance. Additionally, the `not-found.tsx` file was deleted as it is no longer needed in the current setup.